refactor: PostcardRepository 패키지 이동
