+++ src/main/java/com/bookbla/americano/domain/postcard/repository/PostcardRepository.java
@@ -1,7 +1,7 @@
-package com.bookbla.americano.domain.member.repository;
+package com.bookbla.americano.domain.postcard.repository;
 
-import com.bookbla.americano.domain.postcard.Postcard;
-import com.bookbla.americano.domain.postcard.PostcardStatus;
+import com.bookbla.americano.domain.postcard.repository.entity.Postcard;
+import com.bookbla.americano.domain.postcard.repository.entity.PostcardStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostcardRepository extends JpaRepository<Postcard, Long> {
