+++ src/main/java/com/bookbla/americano/domain/postcard/repository/entity/PostcardStatus.java
@@ -1,4 +1,4 @@
-package com.bookbla.americano.domain.postcard;
+package com.bookbla.americano.domain.postcard.repository.entity;
 
 public enum PostcardStatus {
     PENDING, ACCEPT, REFUSED, EXPIRED, ALL_WRONG
